requirements.txt
aiohttp==3.12.4
ruamel.yaml
packaging